# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SMRITI-OS - Sovereign Retail OS",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/customers/": {
            "post": {
                "tags": [
                    "customer"
                ],
                "summary": "Create Customer",
                "description": "Register a new customer with auto-generated code and GSTIN validation.",
                "operationId": "create_customer_api_v1_customers__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/customers/lookup": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Lookup Customer",
                "description": "POS Hot Path: Quick lookup by phone from Shoper9 Customers + Mailinglist.\nReturns loyalty and outstanding balance.",
                "operationId": "lookup_customer_api_v1_customers_lookup_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 10,
                            "title": "Phone"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customers/{customer_id}/ledger": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Get Customer Ledger",
                "description": "Returns the account statement for a customer.",
                "operationId": "get_customer_ledger_api_v1_customers__customer_id__ledger_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Customer Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LedgerEntryResponse"
                                    },
                                    "title": "Response Get Customer Ledger Api V1 Customers  Customer Id  Ledger Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customers/{customer_id}/loyalty/redeem": {
            "post": {
                "tags": [
                    "customer"
                ],
                "summary": "Redeem Loyalty",
                "description": "Redeem loyalty points.\nValidation: \n1. Points available balance check.\n2. Max 50% of invoice value redemption rule.",
                "operationId": "redeem_loyalty_api_v1_customers__customer_id__loyalty_redeem_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Customer Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoyaltyRedeemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customers/search": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Search Customers",
                "description": "Institutional search for customers from Shoper9 by name, phone, or code.\nReturns a list of matching customer profiles.",
                "operationId": "search_customers_api_v1_customers_search_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2,
                            "title": "Q"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customers/{code}": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Get Customer By Code",
                "description": "Fetch a single customer from Shoper9 by their institutional code.",
                "operationId": "get_customer_by_code_api_v1_customers__code__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Code"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomerCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "example": "Amit Kumar"
                    },
                    "phone": {
                        "type": "string",
                        "minLength": 10,
                        "title": "Phone",
                        "example": "9876543210"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "email"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "gstin": {
                        "anyOf": [
                            {
                                "type": "string",
                                "pattern": "^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z]{1}[1-9A-Z]{1}Z[0-9A-Z]{1}$"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gstin"
                    },
                    "address_line1": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address Line1"
                    },
                    "address_line2": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address Line2"
                    },
                    "city": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "City"
                    },
                    "state_code": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 2,
                                "minLength": 2
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "State Code"
                    },
                    "pincode": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pincode"
                    },
                    "credit_limit_paise": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Credit Limit Paise",
                        "default": 0
                    },
                    "credit_days": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Credit Days",
                        "default": 0
                    },
                    "price_group_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Price Group Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "phone"
                ],
                "title": "CustomerCreate"
            },
            "CustomerResponse": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "example": "Amit Kumar"
                    },
                    "phone": {
                        "type": "string",
                        "minLength": 10,
                        "title": "Phone",
                        "example": "9876543210"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "email"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "gstin": {
                        "anyOf": [
                            {
                                "type": "string",
                                "pattern": "^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z]{1}[1-9A-Z]{1}Z[0-9A-Z]{1}$"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gstin"
                    },
                    "address_line1": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address Line1"
                    },
                    "address_line2": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Address Line2"
                    },
                    "city": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "City"
                    },
                    "state_code": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 2,
                                "minLength": 2
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "State Code"
                    },
                    "pincode": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pincode"
                    },
                    "credit_limit_paise": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Credit Limit Paise",
                        "default": 0
                    },
                    "credit_days": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Credit Days",
                        "default": 0
                    },
                    "price_group_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Price Group Id"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "code": {
                        "type": "string",
                        "title": "Code"
                    },
                    "loyalty_points": {
                        "type": "integer",
                        "title": "Loyalty Points"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "outstanding_paise": {
                        "type": "integer",
                        "title": "Outstanding Paise",
                        "default": 0
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "phone",
                    "id",
                    "code",
                    "loyalty_points",
                    "is_active",
                    "created_at"
                ],
                "title": "CustomerResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LedgerEntryResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "txn_type": {
                        "type": "string",
                        "title": "Txn Type"
                    },
                    "txn_ref": {
                        "type": "string",
                        "title": "Txn Ref"
                    },
                    "amount_paise": {
                        "type": "integer",
                        "title": "Amount Paise"
                    },
                    "balance_paise": {
                        "type": "integer",
                        "title": "Balance Paise"
                    },
                    "txn_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Txn Date"
                    },
                    "notes": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notes"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "txn_type",
                    "txn_ref",
                    "amount_paise",
                    "balance_paise",
                    "txn_date",
                    "created_at"
                ],
                "title": "LedgerEntryResponse"
            },
            "LoyaltyRedeemRequest": {
                "properties": {
                    "points_to_redeem": {
                        "type": "integer",
                        "exclusiveMinimum": 0.0,
                        "title": "Points To Redeem"
                    },
                    "invoice_value_paise": {
                        "type": "integer",
                        "exclusiveMinimum": 0.0,
                        "title": "Invoice Value Paise"
                    },
                    "sale_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sale Id"
                    }
                },
                "type": "object",
                "required": [
                    "points_to_redeem",
                    "invoice_value_paise"
                ],
                "title": "LoyaltyRedeemRequest"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}